conn: reorganise
- move functions
- move prototypes
- move struct
- move globals
