+++ conn/conn_globals.h
@@ -33,10 +33,17 @@ extern const char *C_CertificateFile;
 extern const char *C_EntropyFile;
 extern const char *C_SslCiphers;
 extern const char *C_SslClientCert;
+extern bool C_SslUseSslv3;
+extern bool C_SslUseTlsv1;
+extern bool C_SslUseTlsv11;
+extern bool C_SslUseTlsv12;
+extern bool C_SslUseTlsv13;
+extern bool C_SslVerifyDates;
+extern bool C_SslVerifyHost;
+#endif
 #ifdef USE_SSL_GNUTLS
 extern const char *C_SslCaCertificatesFile;
 extern short C_SslMinDhPrimeBits;
-#endif
 #endif
 
 #ifdef USE_SOCKET
@@ -47,16 +54,6 @@ extern const char *C_Tunnel;
 /* These Config Variables are only used in conn/conn_raw.c */
 #ifdef HAVE_GETADDRINFO
 extern bool C_UseIpv6;
-#endif
-
-#ifdef USE_SSL
-extern bool C_SslUseSslv3;
-extern bool C_SslUseTlsv1;
-extern bool C_SslUseTlsv11;
-extern bool C_SslUseTlsv12;
-extern bool C_SslUseTlsv13;
-extern bool C_SslVerifyDates;
-extern bool C_SslVerifyHost;
 #endif
 
 #endif /* MUTT_CONN_CONN_GLOBALS_H */
