+++ conn/socket.h
@@ -50,10 +50,4 @@ int mutt_socket_readchar(struct Connection *conn, char *c);
 int mutt_socket_readln_d(char *buf, size_t buflen, struct Connection *conn, int dbg);
 int mutt_socket_write_d(struct Connection *conn, const char *buf, int len, int dbg);
 
-int raw_socket_read(struct Connection *conn, char *buf, size_t len);
-int raw_socket_write(struct Connection *conn, const char *buf, size_t count);
-int raw_socket_open(struct Connection *conn);
-int raw_socket_close(struct Connection *conn);
-int raw_socket_poll(struct Connection *conn, time_t wait_secs);
-
 #endif /* MUTT_CONN_SOCKET_H */
